more structural changes
