+++ firefly/src/components/game/CodeBlock.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const CodeBlock = () => {
+  return <div />;
+};
+
+export default CodeBlock;
