+++ firefly/src/components/game/FFbox.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const FFbox = () => {
+  return <div />;
+};
+
+export default FFbox;
